Added API Route Protections (#138)
* fixed??

* asd

* Fix for home page and pack-lock clean up

* Added API Route protections

* Added protection for Dashboard

* Added API Route Protections

+++ app/api/rentalItems/[id]/route.ts
@@ -1,6 +1,8 @@
 import { NextRequest, NextResponse } from "next/server";
 import dbConnect from "@/app/lib/dbConnect";
 import RentalItem from "@/app/models/RentalItem";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../../auth/[...nextauth]/options";
 
 export async function GET(
   request: NextRequest,
@@ -27,6 +29,11 @@ export async function PUT(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+  }
+
   try {
     await dbConnect();
     const body = await request.json();
@@ -70,6 +77,12 @@ export async function DELETE(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  //protect
+  const session = await getServerSession(authOptions);
+  if (!session) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+  }
+
   try {
     await dbConnect();
     const deletedItem = await RentalItem.findByIdAndDelete(params.id);
